login

+++ app/Http/Controllers/TasklistsController.php
@@ -15,11 +15,20 @@ class TasklistsController extends Controller
      */
     public function index()
     {
-        $tasklists = Tasklist::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasklists.index', [
-            'tasklists' => $tasklists,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasklists' => $tasklists,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -48,11 +57,11 @@ class TasklistsController extends Controller
             'status' => 'required|max:10',   // 追加
             'content' => 'required|max:191',
         ]);
+        $request->user()->tasklists()->create([
+             'status' => $request->status,
+             'content' => $request->content,
+        ]);
         
-        $tasklist = new Tasklist;
-        $tasklist->status = $request->status; //追加
-        $tasklist->content = $request->content;
-        $tasklist->save();
 
         return redirect('/');
     }
@@ -116,9 +125,13 @@ class TasklistsController extends Controller
      */
     public function destroy($id)
     {
-        $tasklist = Tasklist::find($id);
-        $tasklist->delete();
+        $tasklist = \App\Tasklist::find($id);
+
+        if (\Auth::id() === $tasklist->user_id) {
+            $tasklist->delete();
+        }
 
         return redirect('/');
+ 
     }
 }
